+++ package.json
@@ -30,6 +30,7 @@
     "react-dom": "18.2.0",
     "react-query": "^3.39.3",
     "react-router-dom": "^5.3.0",
+    "tailwind-scrollbar-hide": "^1.1.7",
     "typescript": "4.9.4"
   },
   "devDependencies": {
