+++ tailwind.config.js
@@ -3,11 +3,14 @@ module.exports = {
   content: ["./pages/**/*.{js,ts,jsx,tsx}", "./src/**/*.{js,ts,jsx,tsx}"],
   theme: {
     extend: {
+      colors: {
+        red: "#f96062",
+      },
       maxHeight: {
         common: "50rem",
       },
       margin: {
-        common: "4rem 0 0 0",
+        common: "6rem 0 0 0",
       },
       fontSize: {
         xxs: "10px",
@@ -16,10 +19,13 @@ module.exports = {
         xs: "500px",
         xxs: "360px",
       },
+      boxShadow: {
+        innerCustom: "inset 2px 10px 10px rgba(155,0,0,0.3)",
+      },
     },
   },
   variants: {
     extend: {},
   },
-  plugins: [],
+  plugins: [require("tailwind-scrollbar-hide")],
 };
